@@ -19,11 +19,10 @@ classifiers = [
 Changelog = "https://github.com/thegridelectric/gridworks-protocol/releases"
 
 [tool.poetry.dependencies]
-python = "^3.10"
-pydantic = "^1.10.11"
+python = "^3.11"
+pydantic = "^2.8.2"
 yarl = "^1.9.2"
 pytz = "^2024.1"
-fastapi-utils = "^0.2.1"
 pendulum = "2.1.2"
 ruff = "^0.5.7"
 
@@ -45,6 +44,7 @@ xdoctest = {extras = ["colors"], version = ">=0.15.10"}
 myst-parser = {version = ">=0.16.1"}
 types-pytz = ">=2022.4.0.0"
 rich = ">=12.6.0"
+bump-pydantic = "^0.8.0"
 
 [tool.coverage.paths]
 source = ["src", "*/site-packages"]
@@ -115,7 +115,7 @@ exclude = [
     "dist",
 ]
 output-format = "concise"
-fix = false
+fix = true
 show-fixes = true
 
 [tool.ruff.lint]
@@ -125,13 +125,17 @@ select = ["ALL"]
 ignore = [
     "B027",
     "COM812",
+    "CPY",
     "D",
+    "DOC",
+    "E302",
     "E501",
     "EM",
     "FA", # We only support Python >= 3.10, so we shouldn't need this
     "ISC001",
     "PLR0904",
     "PLW1514",
+    "TRY003", # Suppress for now; lots of them generated by code generation.
     "W191",
 ]
 
