@@ -1,8 +1,11 @@
+# ruff: noqa: ANN401
+
 import abc
 import inspect
 import json
 import re
 import sys
+import typing
 from abc import abstractmethod
 from dataclasses import dataclass
 from typing import (
@@ -19,7 +22,7 @@ from typing import (
 )
 
 import pydantic
-from pydantic import BaseModel, Field, create_model
+from pydantic import BaseModel, Field, ValidationError, create_model
 
 from gwproto.message import Header, Message
 from gwproto.messages import AnyEvent
@@ -39,9 +42,9 @@ class Decoder(abc.ABC):
 class CallableDecoder(Decoder):
     _decode_obj: Callable[[Any], Any]
 
-    def __init__(self, decode_obj: Callable[[Any], Any]):
+    def __init__(self, decode_obj: Callable[[Any], Any]) -> None:
         if not callable(decode_obj):
-            raise ValueError(
+            raise TypeError(
                 f"ERROR. decode_obj {decode_obj}/{type(decode_obj)} attribute is not Callable"
             )
         self._decode_obj = decode_obj
@@ -51,14 +54,14 @@ class CallableDecoder(Decoder):
 
 
 class PydanticDecoder(CallableDecoder):
-    def __init__(self, model: Type["BaseModel"]):
-        super().__init__(model.parse_obj)
+    def __init__(self, model: Type["BaseModel"]) -> None:
+        super().__init__(model.model_validate)
 
 
 class MakerDecoder(CallableDecoder):
     DECODER_FUNCTION_NAME: str = "dict_to_tuple"
 
-    def __init__(self, model: Any):
+    def __init__(self, model: Any) -> None:
         decoder_function = getattr(model, self.DECODER_FUNCTION_NAME, None)
         if decoder_function is None:
             raise ValueError(
@@ -78,50 +81,48 @@ class MessageDecoder(Decoder):
         self,
         decoders: "Decoders",
         message_payload_discriminator: Optional[Type[MessageDiscriminator]] = None,
-    ):
+    ) -> None:
         self.decoders = decoders
         self.message_payload_discriminator = message_payload_discriminator
 
+    @classmethod
+    def _try_message_as_event(cls, message_dict: dict, e: ValidationError) -> Message:
+        special_typename_handling_message: Optional[Message[Any]] = None
+        for error in e.errors():
+            if error.get("type", "") == "union_tag_invalid":
+                ctx = error.get("ctx", {})
+                if ctx.get("discriminator", "") == "'TypeName'" and ctx.get(
+                    "tag", ""
+                ).startswith("gridworks.event"):
+                    try:
+                        message_dict["Payload"] = AnyEvent(**message_dict["Payload"])
+                        special_typename_handling_message = Message(**message_dict)
+                    except Exception as e2:  # noqa: BLE001
+                        raise e2 from e
+        if special_typename_handling_message is None:
+            raise e
+        return special_typename_handling_message
+
     def decode_obj(self, o: Any) -> Message[Any]:
         message_dict: dict = dict(o)
-        message_dict["Header"] = Header.parse_obj(message_dict.get("Header", dict()))
+        message_dict["Header"] = Header.model_validate(message_dict.get("Header", {}))
         message: Message[Any]
         if message_dict["Header"].MessageType in self.decoders:
             message_dict["Payload"] = self.decoders.decode_obj(
                 message_dict["Header"].MessageType,
-                message_dict.get("Payload", dict()),
+                message_dict.get("Payload", {}),
             )
             message = Message(**message_dict)
         else:
             try:
-                message = self.message_payload_discriminator.parse_obj(message_dict)
+                message = self.message_payload_discriminator.model_validate(
+                    message_dict
+                )
             except pydantic.ValidationError as e:
                 # This can result because we receive a TypeName we don't recognize, either because the sender is
                 # newer or older than our code. In some cases we can still meaningfully interpret the message,
                 # for example if we can recognize that it is an event messasge, as is done here.
-                special_typename_handling_message: Optional[Message[Any]] = None
-                for error in e.errors():
-                    if (
-                        error.get("type", "")
-                        == "value_error.discriminated_union.invalid_discriminator"
-                    ):
-                        ctx = error.get("ctx", dict())
-                        if ctx.get("discriminator_key", "") == "TypeName":
-                            if ctx.get("discriminator_value", "").startswith(
-                                "gridworks.event"
-                            ):
-                                try:
-                                    message_dict["Payload"] = AnyEvent(
-                                        **message_dict["Payload"]
-                                    )
-                                    special_typename_handling_message = Message(
-                                        **message_dict
-                                    )
-                                except Exception as e2:
-                                    raise e2 from e
-                if special_typename_handling_message is None:
-                    raise e
-                message = special_typename_handling_message
+                message = self._try_message_as_event(message_dict, e)
         return message
 
 
@@ -130,7 +131,9 @@ class DecoderItem(NamedTuple):
     decoder: Decoder
 
 
-DEFAULT_TYPE_NAME_FIELD = "TypeName"
+DEFAULT_TYPE_NAME_FIELD: str = "TypeName"
+DEFAULT_PAYLOAD_FIELD: str = "Payload"
+DEFAULT_EXCLUDED_TYPE_NAMES: set[str] = {Message.type_name()}
 
 
 @dataclass
@@ -139,8 +142,8 @@ class OneDecoderExtractor:
     decoder_function_name: str = "parse_obj"
 
     def get_type_name(self, obj: Any) -> str:
-        if not (type_name := getattr(obj, self.type_name_field, "")):
-            if (decoder_fields := getattr(obj, "__fields__", None)) is not None:
+        if not (type_name := getattr(obj, self.type_name_field, "")):  # noqa: SIM102
+            if (decoder_fields := getattr(obj, "model_fields", None)) is not None:  # noqa: SIM102
                 if (
                     model_field := decoder_fields.get(self.type_name_field, None)
                 ) is not None:
@@ -157,7 +160,7 @@ class OneDecoderExtractor:
 
     def extract(self, obj: Any) -> Optional[DecoderItem]:
         item = None
-        if type_name := self.get_type_name(obj):
+        if type_name := self.get_type_name(obj):  # noqa: SIM102
             if (decoder := self.get_decoder(obj)) is not None:
                 item = DecoderItem(type_name, decoder)
         return item
@@ -172,15 +175,17 @@ class MakerExtractor(OneDecoderExtractor):
 class Decoders:
     _decoders: dict[str, Decoder]
 
-    def __init__(self, decoders: Optional[dict[str, Decoder]] = None):
-        self._decoders = dict()
+    def __init__(self, decoders: Optional[dict[str, Decoder]] = None) -> None:
+        self._decoders = {}
         if decoders is not None:
             self.add_decoders(decoders)
 
     def decoder(self, type_name: str) -> Decoder:
         return self._decoders[type_name]
 
-    def decode_str(self, type_name: str, content: str | bytes, encoding="utf-8") -> Any:
+    def decode_str(
+        self, type_name: str, content: str | bytes, encoding: str = "utf-8"
+    ) -> Any:
         return self.decoder(type_name).decode_str(content, encoding=encoding)
 
     def decode_obj(self, type_name: str, o: Any) -> Any:
@@ -198,7 +203,7 @@ class Decoders:
         return self
 
     def merge(self, other: "Decoders") -> "Decoders":
-        self.add_decoders(other._decoders)
+        self.add_decoders(other._decoders)  # noqa: SLF001
         return self
 
     def __contains__(self, type_name: str) -> bool:
@@ -208,11 +213,10 @@ class Decoders:
         return list(self._decoders.keys())
 
     def _validate(self, type_name: str, decoder: Decoder) -> None:
-        if type_name in self._decoders:
-            if self._decoders[type_name] is not decoder:
-                raise ValueError(
-                    f"ERROR. decoder for [{type_name}] is already present as [{self._decoders[type_name]}]"
-                )
+        if type_name in self._decoders and self._decoders[type_name] is not decoder:
+            raise ValueError(
+                f"ERROR. decoder for [{type_name}] is already present as [{self._decoders[type_name]}]"
+            )
 
     @classmethod
     def from_objects(
@@ -221,7 +225,7 @@ class Decoders:
         message_payload_discriminator: Optional[Type[MessageDiscriminator]] = None,
         extractors: Optional[Sequence[OneDecoderExtractor]] = None,
     ) -> "Decoders":
-        items = dict()
+        items = {}
         if objs is not None:
             if extractors is None:
                 extractors = [OneDecoderExtractor(), MakerExtractor()]
@@ -242,15 +246,15 @@ class MQTTCodec(abc.ABC):
     ENCODING = "utf-8"
     decoders: Decoders
 
-    def __init__(self, decoders: Decoders):
+    def __init__(self, decoders: Decoders) -> None:
         self.decoders = Decoders().merge(decoders)
         super().__init__()
 
-    def encode(self, content: Any) -> bytes:
+    def encode(self, content: bytes | BaseModel) -> bytes:
         if isinstance(content, bytes):
             encoded = content
         else:
-            encoded = content.json()
+            encoded = content.model_dump_json()
             if not isinstance(encoded, bytes):
                 encoded = encoded.encode(self.ENCODING)
         return encoded
@@ -258,7 +262,7 @@ class MQTTCodec(abc.ABC):
     def decode(self, topic: str, payload: bytes) -> Any:
         decoded_topic = MQTTTopic.decode(topic)
         if decoded_topic.envelope_type not in self.decoders:
-            raise Exception(
+            raise ValueError(
                 f"Type {decoded_topic.envelope_type} not recognized. Available decoders: {self.decoders.types()}"
             )
         self.validate_source_alias(decoded_topic.src)
@@ -267,25 +271,30 @@ class MQTTCodec(abc.ABC):
         )
 
     @abstractmethod
-    def validate_source_alias(self, source_alias: str): ...
+    def validate_source_alias(self, source_alias: str) -> None: ...
 
 
 def get_pydantic_literal_type_name(
     o: Any, type_name_field: str = DEFAULT_TYPE_NAME_FIELD
 ) -> str:
-    if hasattr(o, "__fields__"):
-        if type_name_field in o.__fields__:
-            if get_origin(o.__fields__[type_name_field].annotation) == Literal:
-                return str(o.__fields__[type_name_field].default)
+    if (
+        hasattr(o, "model_fields")
+        and type_name_field in o.model_fields
+        and get_origin(o.model_fields[type_name_field].annotation) == Literal
+    ):
+        return str(o.model_fields[type_name_field].default)
     return ""
 
 
-def pydantic_named_types(
+def pydantic_named_types(  # noqa: C901
     module_names: str | Sequence[str],
     modules: Optional[Sequence[Any]] = None,
     type_name_field: str = DEFAULT_TYPE_NAME_FIELD,
     type_name_regex: Optional[re.Pattern] = None,
+    excluded_type_names: Optional[set[str]] = None,
 ) -> list[Any]:
+    if excluded_type_names is None:
+        excluded_type_names = DEFAULT_EXCLUDED_TYPE_NAMES
     if isinstance(module_names, str):
         module_names = [module_names]
     if unimported := [
@@ -293,7 +302,7 @@ def pydantic_named_types(
     ]:
         raise ValueError(f"ERROR. modules {unimported} have not been imported.")
     named_types = []
-    type_names: dict[str, Any] = dict()
+    accumulated_types: dict[str, Any] = {}
     if modules is None:
         modules = []
     for module in [sys.modules[module_name] for module_name in module_names] + list(
@@ -306,41 +315,44 @@ def pydantic_named_types(
             if type_name := get_pydantic_literal_type_name(
                 module_class, type_name_field=type_name_field
             ):
-                if type_name in type_names:
-                    if type_names[type_name] is module_class:
+                if type_name in excluded_type_names:
+                    continue
+                if type_name in accumulated_types:
+                    if accumulated_types[type_name] is module_class:
                         continue
                     raise ValueError(
                         f"ERROR {type_name_field} ({type_name}) "
                         f"for {module_class} already seen for "
-                        f"class {type_names[type_name]}"
+                        f"class {accumulated_types[type_name]}"
                     )
-                if type_name_regex is not None:
-                    if not type_name_regex.match(type_name):
-                        continue
-                type_names[type_name] = module_class
+                if type_name_regex is not None and not type_name_regex.match(type_name):
+                    continue
+                accumulated_types[type_name] = module_class
                 named_types.append(module_class)
     return named_types
 
 
-def create_message_payload_discriminator(
+def create_message_payload_discriminator(  # noqa: PLR0913, PLR0917, RUF100
     model_name: str,
     module_names: str | Sequence[str] = "",
     modules: Optional[Sequence[Any]] = None,
     explicit_types: Optional[Sequence[Any]] = None,
     type_name_field: str = DEFAULT_TYPE_NAME_FIELD,
     type_name_regex: Optional[re.Pattern] = None,
+    excluded_type_names: Optional[set[str]] = None,
 ) -> Type[MessageDiscriminator]:
     used_types = pydantic_named_types(
         module_names=module_names,
         modules=modules,
         type_name_field=type_name_field,
         type_name_regex=type_name_regex,
+        excluded_type_names=excluded_type_names,
     )
     if explicit_types is not None:
         used_types.extend(explicit_types)
     return create_model(
         model_name,
-        __base__=Message,  # type: ignore
+        __base__=Message,
         Payload=(
             Union[tuple(used_types)],
             Field(..., discriminator=type_name_field),
@@ -348,20 +360,22 @@ def create_message_payload_discriminator(
     )
 
 
-def create_discriminator(
+def create_discriminator(  # noqa: PLR0913, PLR0917, RUF100
     model_name: str,
     module_names: str | Sequence[str] = "",
     modules: Optional[Sequence[Any]] = None,
     explicit_types: Optional[Sequence[Any]] = None,
     type_name_field: str = DEFAULT_TYPE_NAME_FIELD,
     type_name_regex: Optional[re.Pattern] = None,
-    payload_field_name="Payload",
+    payload_field_name: str = DEFAULT_PAYLOAD_FIELD,
+    excluded_type_names: Optional[set[str]] = None,
 ) -> BaseModel:
     used_types = pydantic_named_types(
         module_names=module_names,
         modules=modules,
         type_name_field=type_name_field,
         type_name_regex=type_name_regex,
+        excluded_type_names=excluded_type_names,
     )
     if explicit_types is not None:
         used_types.extend(explicit_types)
@@ -375,8 +389,7 @@ def create_discriminator(
             Union[tuple(used_types)],
             Field(..., discriminator=type_name_field),
         )
-    payload_field_kwargs = {payload_field_name: payload_field_type}
-    return create_model(model_name, **payload_field_kwargs)
+    return create_model(model_name, **{payload_field_name: payload_field_type})
 
 
 class PydanticTypeNameDecoder(Decoder):
@@ -384,7 +397,7 @@ class PydanticTypeNameDecoder(Decoder):
     payload_field_name: str
     contains: set[str]
 
-    def __init__(
+    def __init__(  # noqa: PLR0913
         self,
         model_name: str,
         *,
@@ -393,8 +406,8 @@ class PydanticTypeNameDecoder(Decoder):
         explicit_types: Optional[Sequence[Any]] = None,
         type_name_field: str = DEFAULT_TYPE_NAME_FIELD,
         type_name_regex: Optional[re.Pattern] = None,
-        payload_field_name="Payload",
-    ):
+        payload_field_name: str = "Payload",
+    ) -> None:
         self.payload_field_name = payload_field_name
         self.loader = create_discriminator(
             model_name=model_name,
@@ -405,17 +418,23 @@ class PydanticTypeNameDecoder(Decoder):
             type_name_regex=type_name_regex,
             payload_field_name=payload_field_name,
         )
-        payload_field = self.loader.__fields__[self.payload_field_name]
-        if payload_field.sub_fields_mapping is not None:
-            self.contains = set(payload_field.sub_fields_mapping.keys())
-        else:
-            self.contains = {payload_field.type_.__fields__["TypeName"].default}
+        payload_field_annotation = self.loader.model_fields[
+            self.payload_field_name
+        ].annotation
+        if payload_field_annotation is None:
+            raise ValueError(
+                f"ERROR. Payload field <{payload_field_name}> has no annotation"
+            )
+        self.contains = {
+            payload_type.model_fields["TypeName"].default
+            for payload_type in typing.get_args(payload_field_annotation)
+        }
 
     def __contains__(self, type_name: str) -> bool:
         return type_name in self.contains
 
     def decode_obj(self, data: dict) -> BaseModel:
         return getattr(
-            self.loader.parse_obj({self.payload_field_name: data}),
+            self.loader.model_validate({self.payload_field_name: data}),
             self.payload_field_name,
         )
