@@ -4,7 +4,7 @@ import json
 import logging
 from typing import Any, Dict, Literal, Optional
 
-from pydantic import BaseModel, Field, validator
+from pydantic import BaseModel, Field, field_validator
 
 from gwproto.data_classes.component import Component
 from gwproto.errors import SchemaError
@@ -60,7 +60,8 @@ class ComponentGt(BaseModel):
     TypeName: Literal["component.gt"] = "component.gt"
     Version: Literal["000"] = "000"
 
-    @validator("ComponentId")
+    @field_validator("ComponentId")
+    @classmethod
     def _check_component_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -70,7 +71,8 @@ class ComponentGt(BaseModel):
             )
         return v
 
-    @validator("ComponentAttributeClassId")
+    @field_validator("ComponentAttributeClassId")
+    @classmethod
     def _check_component_attribute_class_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -98,8 +100,8 @@ class ComponentGt(BaseModel):
         """
         d = {
             key: value
-            for key, value in self.dict(
-                include=self.__fields_set__ | {"TypeName", "Version"},
+            for key, value in self.model_dump(
+                include=self.model_fields_set | {"TypeName", "Version"},
                 by_alias=True,
             ).items()
             if value is not None
