@@ -2,9 +2,9 @@
 
 import json
 import logging
-from typing import Any, Dict, List, Literal, Optional
+from typing import Any, Dict, List, Literal, Optional, Self
 
-from pydantic import BaseModel, Field, root_validator, validator
+from pydantic import BaseModel, Field, field_validator, model_validator
 
 from gwproto.data_classes.components.electric_meter_component import (
     ElectricMeterComponent,
@@ -93,7 +93,8 @@ class ElectricMeterComponentGt(BaseModel):
     TypeName: Literal["electric.meter.component.gt"] = "electric.meter.component.gt"
     Version: Literal["000"] = "000"
 
-    @validator("ComponentId")
+    @field_validator("ComponentId")
+    @classmethod
     def _check_component_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -103,7 +104,8 @@ class ElectricMeterComponentGt(BaseModel):
             )
         return v
 
-    @validator("ComponentAttributeClassId")
+    @field_validator("ComponentAttributeClassId")
+    @classmethod
     def _check_component_attribute_class_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -113,7 +115,8 @@ class ElectricMeterComponentGt(BaseModel):
             )
         return v
 
-    @validator("ModbusPort")
+    @field_validator("ModbusPort")
+    @classmethod
     def _check_modbus_port(cls, v: Optional[int]) -> Optional[int]:
         if v is None:
             return v
@@ -125,46 +128,36 @@ class ElectricMeterComponentGt(BaseModel):
             )
         return v
 
-    @root_validator
-    def check_axiom_1(cls, v: dict) -> dict:
+    @model_validator(mode="after")
+    def check_axiom_1(self) -> Self:
         """
         Axiom 1: Modbus consistency.
-        ModbusHost is None if and only if ModbusPort is None
+        ModbusHost requires a ModbusPort
         """
-        # TODO: Implement check for axiom 1"
-        ModbusHost = v.get("ModbusHost")
-        ModbusPort = v.get("ModbusHost")
-        if ModbusHost is None and ModbusPort is not None:
+        if self.ModbusHost is not None and self.ModbusPort is None:
             raise ValueError("Axiom 1: ModbusHost None iff ModbusPort None! ")
-        if ModbusHost is not None and ModbusPort is None:
-            raise ValueError("Axiom 1: ModbusHost None iff ModbusPort None! ")
-        return v
+        return self
 
-    @root_validator
-    def check_axiom_2(cls, v: dict) -> dict:
+    @model_validator(mode="after")
+    def check_axiom_2(self) -> Self:
         """
         Axiom 2: Egauge4030 consistency.
         If the EgaugeIoList has non-zero length, then the ModbusHost is not None and
         the set of output configs is equal to ConfigList as a set
         """
-        # TODO: Implement check for axiom 2"
-        EgaugeIoList = v.get("EgaugeIoList")
-        ModbusHost = v.get("ModbusHost")
-        ConfigList = v.get("ConfigList")
-        if len(EgaugeIoList) == 0:
-            return v
+        if len(self.EgaugeIoList) == 0:
+            return self
 
-        if ModbusHost is None:
+        if self.ModbusHost is None:
             raise ValueError(
                 "Axiom 2: If EgaugeIoList has non-zero length then ModbusHost must exist!"
             )
-        output_configs = set(map(lambda x: x.OutputConfig, EgaugeIoList))
-        if output_configs != set(ConfigList):
+        if {x.OutputConfig for x in self.EgaugeIoList} != set(self.ConfigList):
             raise ValueError(
-                "Axiom 2: If EgaugeIoList has non-zero length then then the set of"
+                "Axiom 2: If EgaugeIoList has non-zero length then the set of"
                 "output configs must equal ConfigList as a set"
             )
-        return v
+        return self
 
     def as_dict(self) -> Dict[str, Any]:
         """
@@ -184,8 +177,8 @@ class ElectricMeterComponentGt(BaseModel):
         """
         d = {
             key: value
-            for key, value in self.dict(
-                include=self.__fields_set__ | {"TypeName", "Version"}
+            for key, value in self.model_dump(
+                include=self.model_fields_set | {"TypeName", "Version"}
             ).items()
             if value is not None
         }
