@@ -4,7 +4,7 @@ import json
 import logging
 from typing import Any, Dict, List, Literal
 
-from pydantic import BaseModel, Field, root_validator, validator
+from pydantic import BaseModel, Field, field_validator, model_validator
 
 from gwproto.enums import TelemetryName as EnumTelemetryName
 from gwproto.errors import SchemaError
@@ -53,7 +53,8 @@ class GtShSimpleTelemetryStatus(BaseModel):
     TypeName: Literal["gt.sh.simple.telemetry.status"] = "gt.sh.simple.telemetry.status"
     Version: Literal["100"] = "100"
 
-    @validator("ShNodeAlias")
+    @field_validator("ShNodeAlias")
+    @classmethod
     def _check_sh_node_alias(cls, v: str) -> str:
         try:
             check_is_left_right_dot(v)
@@ -61,7 +62,8 @@ class GtShSimpleTelemetryStatus(BaseModel):
             raise ValueError(f"ShNodeAlias failed LeftRightDot format validation: {e}")
         return v
 
-    @validator("ReadTimeUnixMsList")
+    @field_validator("ReadTimeUnixMsList")
+    @classmethod
     def _check_read_time_unix_ms_list(cls, v: List[int]) -> List[int]:
         for elt in v:
             try:
@@ -72,19 +74,17 @@ class GtShSimpleTelemetryStatus(BaseModel):
                 )
         return v
 
-    @root_validator
-    def check_axiom_1(cls, v: dict) -> dict:
+    @model_validator(mode="after")
+    def check_axiom_1(self) -> dict:
         """
         Axiom 1: ListLengthConsistency.
         ValueList and ReadTimeUnixMsList must have the same length.
         """
-        value_list: List[int] = v.get("ValueList")
-        time_list: List[int] = v.get("ReadTimeUnixMsList")
-        if len(value_list) != len(time_list):
+        if len(self.ValueList) != len(self.ReadTimeUnixMsList):
             raise ValueError(
                 "Axiom 1: ValueList and ReadTimeUnixMsList must have the same length."
             )
-        return v
+        return self
 
     def as_dict(self) -> Dict[str, Any]:
         """
@@ -104,8 +104,8 @@ class GtShSimpleTelemetryStatus(BaseModel):
         """
         d = {
             key: value
-            for key, value in self.dict(
-                include=self.__fields_set__ | {"TypeName", "Version"}
+            for key, value in self.model_dump(
+                include=self.model_fields_set | {"TypeName", "Version"}
             ).items()
             if value is not None
         }
