@@ -4,7 +4,7 @@ import json
 import logging
 from typing import Any, Dict, List, Literal
 
-from pydantic import BaseModel, Field, validator
+from pydantic import BaseModel, Field, field_validator
 
 from gwproto.errors import SchemaError
 from gwproto.types.gt_sh_booleanactuator_cmd_status import (
@@ -62,7 +62,8 @@ class GtShStatus(BaseModel):
     TypeName: Literal["gt.sh.status"] = "gt.sh.status"
     Version: Literal["110"] = "110"
 
-    @validator("FromGNodeAlias")
+    @field_validator("FromGNodeAlias")
+    @classmethod
     def _check_from_g_node_alias(cls, v: str) -> str:
         try:
             check_is_left_right_dot(v)
@@ -72,7 +73,8 @@ class GtShStatus(BaseModel):
             )
         return v
 
-    @validator("FromGNodeId")
+    @field_validator("FromGNodeId")
+    @classmethod
     def _check_from_g_node_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -82,7 +84,8 @@ class GtShStatus(BaseModel):
             )
         return v
 
-    @validator("AboutGNodeAlias")
+    @field_validator("AboutGNodeAlias")
+    @classmethod
     def _check_about_g_node_alias(cls, v: str) -> str:
         try:
             check_is_left_right_dot(v)
@@ -92,7 +95,8 @@ class GtShStatus(BaseModel):
             )
         return v
 
-    @validator("SlotStartUnixS")
+    @field_validator("SlotStartUnixS")
+    @classmethod
     def _check_slot_start_unix_s(cls, v: int) -> int:
         try:
             check_is_reasonable_unix_time_s(v)
@@ -102,7 +106,8 @@ class GtShStatus(BaseModel):
             )
         return v
 
-    @validator("StatusUid")
+    @field_validator("StatusUid")
+    @classmethod
     def _check_status_uid(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -130,8 +135,8 @@ class GtShStatus(BaseModel):
         """
         d = {
             key: value
-            for key, value in self.dict(
-                include=self.__fields_set__ | {"TypeName", "Version"}
+            for key, value in self.model_dump(
+                include=self.model_fields_set | {"TypeName", "Version"}
             ).items()
             if value is not None
         }
