@@ -4,7 +4,8 @@ from functools import cached_property
 from typing import Optional
 
 import yarl
-from pydantic import BaseModel, Extra, conint, validator
+from pydantic import BaseModel, ConfigDict, Field, field_validator
+from typing_extensions import Annotated
 
 from gwproto.enums import TelemetryName, Unit
 from gwproto.types.hubitat_component_gt import HubitatRESTResolutionSettings
@@ -26,10 +27,10 @@ HUBITAT_ACCESS_TOKEN_REGEX = re.compile(
 
 
 class FibaroTempSensorSettingsGt(BaseModel):
-    stack_depth: conint(ge=1)
+    stack_depth: Annotated[int, Field(ge=1)]
     device_id: int
     fibaro_component_id: str
-    analog_input_id: conint(ge=1, le=2)
+    analog_input_id: Annotated[int, Field(ge=1, le=2)]
     tank_label: str = ""
     exponent: int = 1
     telemetry_name_gt_enum_symbol: str = "c89d0ba1"
@@ -44,19 +45,19 @@ class FibaroTempSensorSettingsGt(BaseModel):
     4. The default value.
     """
     rest: Optional[RESTPollerSettings] = None
+    model_config = ConfigDict(
+        extra="allow", alias_generator=snake_to_camel, populate_by_name=True
+    )
 
-    class Config:
-        extra = Extra.allow
-        alias_generator = snake_to_camel
-        allow_population_by_field_name = True
-
-    @validator("telemetry_name_gt_enum_symbol")
+    @field_validator("telemetry_name_gt_enum_symbol")
+    @classmethod
     def _check_telemetry_name_symbol(cls, v: str) -> str:
         if v not in TelemetryName.symbols():
             v = TelemetryName.value_to_symbol(TelemetryName.default())
         return v
 
-    @validator("temp_unit_gt_enum_symbol")
+    @field_validator("temp_unit_gt_enum_symbol")
+    @classmethod
     def _checktemp_unit_gt_enum_symbol(cls, v: str) -> str:
         if v not in Unit.symbols():
             v = Unit.value_to_symbol(Unit.default())
@@ -68,11 +69,10 @@ DEFAULT_SENSOR_NODE_NAME_FORMAT = "{tank_name}.temp.depth{stack_depth}"
 
 class FibaroTempSensorSettings(FibaroTempSensorSettingsGt):
     node_name: str
+    model_config = ConfigDict(ignored_types=(cached_property, TelemetryName))
 
-    class Config:
-        keep_untouched = (cached_property, TelemetryName)
-
-    @validator("rest")
+    @field_validator("rest")
+    @classmethod
     def _collapse_rest_url(cls, v: Optional[RESTPollerSettings]):
         if v is not None:
             # Collapse session.base_url and request.url into
@@ -216,7 +216,7 @@ class FibaroTempSensorSettings(FibaroTempSensorSettingsGt):
                 tank_name=tank_name,
                 stack_depth=settings_gt.stack_depth,
             ),
-            **settings_gt.dict(),
+            **settings_gt.model_dump(),
         )
         if settings.poll_period_seconds is None:
             settings.poll_period_seconds = default_poll_period_seconds
@@ -233,8 +233,6 @@ class HubitatTankSettingsGt(BaseModel):
     default_poll_period_seconds: Optional[float] = None
     devices: list[FibaroTempSensorSettingsGt] = []
     web_listen_enabled: bool = True
-
-    class Config:
-        extra = Extra.allow
-        alias_generator = snake_to_camel
-        allow_population_by_field_name = True
+    model_config = ConfigDict(
+        extra="allow", alias_generator=snake_to_camel, populate_by_name=True
+    )
