@@ -4,7 +4,7 @@ import json
 import logging
 from typing import Any, Dict, Literal, Optional
 
-from pydantic import BaseModel, Field, validator
+from pydantic import BaseModel, Field, field_validator
 
 from gwproto.data_classes.cacs.resistive_heater_cac import ResistiveHeaterCac
 from gwproto.enums import MakeModel as EnumMakeModel
@@ -50,7 +50,8 @@ class ResistiveHeaterCacGt(BaseModel):
     TypeName: Literal["resistive.heater.cac.gt"] = "resistive.heater.cac.gt"
     Version: Literal["000"] = "000"
 
-    @validator("ComponentAttributeClassId")
+    @field_validator("ComponentAttributeClassId")
+    @classmethod
     def _check_component_attribute_class_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -60,7 +61,8 @@ class ResistiveHeaterCacGt(BaseModel):
             )
         return v
 
-    @validator("RatedVoltageV")
+    @field_validator("RatedVoltageV")
+    @classmethod
     def _check_rated_voltage_v(cls, v: int) -> int:
         try:
             check_is_positive_integer(v)
@@ -88,8 +90,8 @@ class ResistiveHeaterCacGt(BaseModel):
         """
         d = {
             key: value
-            for key, value in self.dict(
-                include=self.__fields_set__ | {"TypeName", "Version"}
+            for key, value in self.model_dump(
+                include=self.model_fields_set | {"TypeName", "Version"}
             ).items()
             if value is not None
         }
