@@ -2,9 +2,9 @@
 
 import json
 import logging
-from typing import Any, Dict, Literal, Optional
+from typing import Any, Dict, Literal, Optional, Self
 
-from pydantic import BaseModel, Field, root_validator, validator
+from pydantic import BaseModel, Field, field_validator, model_validator
 
 from gwproto.enums import TelemetryName as EnumTelemetryName
 from gwproto.enums import Unit as EnumUnit
@@ -55,7 +55,8 @@ class TelemetryReportingConfig(BaseModel):
     TypeName: Literal["telemetry.reporting.config"] = "telemetry.reporting.config"
     Version: Literal["000"] = "000"
 
-    @validator("AboutNodeName")
+    @field_validator("AboutNodeName")
+    @classmethod
     def _check_about_node_name(cls, v: str) -> str:
         try:
             check_is_left_right_dot(v)
@@ -65,7 +66,8 @@ class TelemetryReportingConfig(BaseModel):
             )
         return v
 
-    @validator("NameplateMaxValue")
+    @field_validator("NameplateMaxValue")
+    @classmethod
     def _check_nameplate_max_value(cls, v: Optional[int]) -> Optional[int]:
         if v is None:
             return v
@@ -77,20 +79,17 @@ class TelemetryReportingConfig(BaseModel):
             )
         return v
 
-    @root_validator
-    def check_axiom_1(cls, v: dict) -> dict:
+    @model_validator(mode="after")
+    def check_axiom_1(self) -> Self:
         """
         Axiom 1: Async reporting consistency.
         If AsyncReportThreshold exists, so does NameplateMaxValue
         """
-        AsyncReportThreshold = v.get("AsyncReportThreshold")
-        NameplateMaxValue = v.get("NameplateMaxValue")
-        if AsyncReportThreshold is not None:
-            if NameplateMaxValue is None:
-                raise ValueError(
-                    "Violates Axiom 1: If AsyncReportThreshold exists, so does NameplateMaxValue"
-                )
-        return v
+        if self.AsyncReportThreshold is not None and self.NameplateMaxValue is None:
+            raise ValueError(
+                "Violates Axiom 1: If AsyncReportThreshold exists, so does NameplateMaxValue"
+            )
+        return self
 
     def as_dict(self) -> Dict[str, Any]:
         """
@@ -110,8 +109,8 @@ class TelemetryReportingConfig(BaseModel):
         """
         d = {
             key: value
-            for key, value in self.dict(
-                include=self.__fields_set__ | {"TypeName", "Version"}
+            for key, value in self.model_dump(
+                include=self.model_fields_set | {"TypeName", "Version"}
             ).items()
             if value is not None
         }
