@@ -1,4 +1,4 @@
-from pydantic import BaseModel, Extra
+from pydantic import BaseModel, ConfigDict
 
 from gwproto.utils import snake_to_camel
 
@@ -11,8 +11,6 @@ class WebServerGt(BaseModel):
     Port: int = 8080
     Enabled: bool = True
     Kwargs: dict = {}
-
-    class Config:
-        extra = Extra.allow
-        alias_generator = snake_to_camel
-        allow_population_by_field_name = True
+    model_config = ConfigDict(
+        extra="allow", alias_generator=snake_to_camel, populate_by_name=True
+    )
