@@ -1,21 +1,12 @@
 import json
 import time
 import uuid
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Any, Optional, Type
 
-from gwproto.messages import AnyEvent
-from pydantic import ValidationError
-
-try:
-    from rich import print
-except ImportError:
-    pass
-
-from gwproto import Message, MQTTCodec
+from gwproto import Message
 from gwproto.messages import (
     Ack,
+    AnyEvent,
     GtDispatchBoolean_Maker,
     GtShCliAtnCmd_Maker,
     GtShStatus_Maker,
@@ -35,17 +26,19 @@ from gwproto.messages import (
     SnapshotSpaceheatEvent,
     StartupEvent,
 )
+from pydantic import ValidationError
 
 from tests.dummy_decoders import CHILD, PARENT
 from tests.dummy_decoders.child.codec import ChildMQTTCodec
 from tests.dummy_decoders.parent.codec import ParentMQTTCodec
 
+from .decode_utils import MessageCase, assert_encode_decode
+
 TEST_DATA_DIR = Path(__file__).parent / "data"
 
 
 def get_stored_message_dicts() -> dict:
-    d = {}
-    return d
+    return {}
 
 
 def child_to_parent_payload_dicts() -> dict:
@@ -57,31 +50,23 @@ def child_to_parent_payload_dicts() -> dict:
     return d
 
 
-@dataclass
-class MessageCase:
-    src_message: Message
-    exp_message: Optional[Message] = None
-    exp_payload: Any = None
-    exp_exceptions: list[Type[Exception]] = field(default_factory=list)
-
-
 def child_to_parent_messages() -> list[MessageCase]:
     stored_message_dicts = child_to_parent_payload_dicts()
     status_message_dict = stored_message_dicts["status"]
     gt_sh_status = GtShStatus_Maker.dict_to_tuple(status_message_dict["Payload"])
     gt_sh_status_event = GtShStatusEvent(Src=CHILD, status=gt_sh_status)
-    unrecognized_status_event = AnyEvent(**gt_sh_status_event.dict())
+    unrecognized_status_event = AnyEvent(**gt_sh_status_event.model_dump())
     unrecognized_status_event.TypeName += ".foo"
     unrecognized_event = AnyEvent(
         TypeName="gridworks.event.bar", MessageId="1", TimeNS=1, Src="1"
     )
     unrecognizeable_not_event_type = AnyEvent(
         **dict(
-            gt_sh_status_event.dict(),
+            gt_sh_status_event.model_dump(),
             TypeName="bla",
         )
     )
-    unrecognizeable_bad_event_content = dict(TypeName="gridworks.event.baz")
+    unrecognizeable_bad_event_content = {"TypeName": "gridworks.event.baz"}
     snap_message_dict = stored_message_dicts["snapshot"]
     snapshot_spaceheat = SnapshotSpaceheat_Maker.dict_to_tuple(
         snap_message_dict["Payload"]
@@ -91,42 +76,56 @@ def child_to_parent_messages() -> list[MessageCase]:
     return [
         # Gs Pwr
         MessageCase(
+            "GsPwr",
             Message(
                 Src=CHILD, MessageType="power.watts", Payload=PowerWatts_Maker(1).tuple
-            )
+            ),
         ),
         # status
         # QUESTION: why does this fail when replacing "gt.sh.status.110" with "gt.sh.status"?
         MessageCase(
+            "status-payload-obj",
             Message(Src=CHILD, MessageType="gt.sh.status.110", Payload=gt_sh_status),
             None,
             gt_sh_status,
         ),
         MessageCase(
+            "status-payload-dict",
             Message(Src=CHILD, Payload=status_message_dict["Payload"]),
             None,
             gt_sh_status,
         ),
         MessageCase(
-            Message(Src=CHILD, Payload=gt_sh_status.as_dict()), None, gt_sh_status
+            "status-payload-as_dict",
+            Message(Src=CHILD, Payload=gt_sh_status.as_dict()),
+            None,
+            gt_sh_status,
         ),
         # snapshot
-        MessageCase(Message(**snap_message_dict), None, snapshot_spaceheat),
+        MessageCase("snap", Message(**snap_message_dict), None, snapshot_spaceheat),
         MessageCase(
+            "snap-payload-dict",
             Message(Src=CHILD, Payload=snap_message_dict["Payload"]),
             None,
             snapshot_spaceheat,
         ),
         MessageCase(
+            "snap-payload-as_dict",
             Message(Src=CHILD, Payload=snapshot_spaceheat.as_dict()),
             None,
             snapshot_spaceheat,
         ),
         # events
-        MessageCase(Message(Src=CHILD, Payload=gt_sh_status_event)),
-        MessageCase(Message(Src=CHILD, Payload=unrecognized_status_event)),
-        MessageCase(Message(Src=CHILD, Payload=unrecognized_event)),
+        MessageCase("event-status", Message(Src=CHILD, Payload=gt_sh_status_event)),
         MessageCase(
+            "event-unrecognized-status",
+            Message(Src=CHILD, Payload=unrecognized_status_event),
+        ),
+        MessageCase(
+            "event-unrecognized", Message(Src=CHILD, Payload=unrecognized_event)
+        ),
+        MessageCase(
+            "unrecognized-not-event",
             Message(
                 Src=CHILD,
                 Payload=unrecognizeable_not_event_type,
@@ -134,34 +133,52 @@ def child_to_parent_messages() -> list[MessageCase]:
             exp_exceptions=[ValidationError],
         ),
         MessageCase(
+            "unrecognizeable-bad-event",
             Message(
                 Src=CHILD,
                 Payload=unrecognizeable_bad_event_content,
             ),
             exp_exceptions=[ValidationError],
         ),
-        MessageCase(Message(Src=CHILD, Payload=snapshot_event)),
-        MessageCase(Message(Src=CHILD, Payload=StartupEvent())),
-        MessageCase(Message(Src=CHILD, Payload=ShutdownEvent(Reason="foo"))),
+        MessageCase("snap-payload-obj", Message(Src=CHILD, Payload=snapshot_event)),
+        MessageCase("startup-event", Message(Src=CHILD, Payload=StartupEvent())),
         MessageCase(
+            "shutdown-event", Message(Src=CHILD, Payload=ShutdownEvent(Reason="foo"))
+        ),
+        MessageCase(
+            "problem-event",
             Message(
                 Src=CHILD,
                 Payload=ProblemEvent(ProblemType=Problems.error, Summary="foo"),
-            )
+            ),
         ),
-        MessageCase(Message(Src=CHILD, Payload=MQTTConnectEvent(PeerName=PARENT))),
         MessageCase(
-            Message(Src=CHILD, Payload=MQTTConnectFailedEvent(PeerName=PARENT))
+            "mqtt-connect-event",
+            Message(Src=CHILD, Payload=MQTTConnectEvent(PeerName=PARENT)),
         ),
-        MessageCase(Message(Src=CHILD, Payload=MQTTDisconnectEvent(PeerName=PARENT))),
         MessageCase(
-            Message(Src=CHILD, Payload=MQTTFullySubscribedEvent(PeerName=PARENT))
+            "mqtt-conenct-failed-event",
+            Message(Src=CHILD, Payload=MQTTConnectFailedEvent(PeerName=PARENT)),
         ),
-        MessageCase(Message(Src=CHILD, Payload=ResponseTimeoutEvent(PeerName=PARENT))),
-        MessageCase(Message(Src=CHILD, Payload=PeerActiveEvent(PeerName=PARENT))),
+        MessageCase(
+            "mqtt-disconnect-event",
+            Message(Src=CHILD, Payload=MQTTDisconnectEvent(PeerName=PARENT)),
+        ),
+        MessageCase(
+            "mqtt-fully-subscribed-event",
+            Message(Src=CHILD, Payload=MQTTFullySubscribedEvent(PeerName=PARENT)),
+        ),
+        MessageCase(
+            "response-timeout-event",
+            Message(Src=CHILD, Payload=ResponseTimeoutEvent(PeerName=PARENT)),
+        ),
+        MessageCase(
+            "peer-active-event",
+            Message(Src=CHILD, Payload=PeerActiveEvent(PeerName=PARENT)),
+        ),
         # misc messages
-        MessageCase(PingMessage(Src=CHILD)),
-        MessageCase(Message(Src=CHILD, Payload=Ack(AckMessageID="1"))),
+        MessageCase("ping", PingMessage(Src=CHILD)),
+        MessageCase("ack", Message(Src=CHILD, Payload=Ack(AckMessageID="1"))),
     ]
 
 
@@ -181,96 +198,25 @@ def parent_to_child_messages() -> list[MessageCase]:
     ).tuple
     return [
         # misc messages
-        MessageCase(PingMessage(Src=PARENT)),
-        MessageCase(Message(Src=PARENT, Payload=Ack(AckMessageID="1"))),
+        MessageCase("ping", PingMessage(Src=PARENT)),
+        MessageCase("ack", Message(Src=PARENT, Payload=Ack(AckMessageID="1"))),
         MessageCase(
+            "snap",
             Message(Src=PARENT, Payload=snapshot_request.as_dict()),
             None,
             snapshot_request,
         ),
-        MessageCase(Message(Src=PARENT, Payload=set_relay.as_dict()), None, set_relay),
+        MessageCase(
+            "set-relay",
+            Message(Src=PARENT, Payload=set_relay.as_dict()),
+            None,
+            set_relay,
+        ),
     ]
 
 
-def assert_encode_decode(
-    src_codec: MQTTCodec,
-    dst_codec: MQTTCodec,
-    messages: list[MessageCase],
-):
-    errors = []
-    for i, case in enumerate(messages):
-        path_dbg = 0
-        # old_len = len(errors)  # noqa: ERA001
-        try:
-            decoded = dst_codec.decode(
-                case.src_message.mqtt_topic(),
-                src_codec.encode(case.src_message),
-            )
-        except Exception as e:
-            path_dbg |= 0x00000001
-            if type(e) in case.exp_exceptions:
-                path_dbg |= 0x00000002
-                continue
-            else:
-                path_dbg |= 0x00000004
-                errors.append(i)
-                if len(errors) == 1:
-                    path_dbg |= 0x00000008
-                    print(f"FIRST ERROR, at index {i}")
-                    print(f"exp expected exception in {case.exp_exceptions}")
-                    print(f"got: <{type(e)}> <{e}>")
-
-        else:
-            path_dbg |= 0x00000010
-            if case.exp_exceptions:
-                path_dbg |= 0x00000020
-                errors.append(i)
-                if len(errors) == 1:
-                    path_dbg |= 0x00000040
-                    print(f"FIRST ERROR, at index {i}")
-                    print(f"exp expected exception in {case.exp_exceptions}")
-                    print(f"got: {decoded}")
-            else:
-                path_dbg |= 0x00000080
-                decoded = dst_codec.decode(
-                    case.src_message.mqtt_topic(),
-                    src_codec.encode(case.src_message),
-                )
-                if case.exp_message is not None:
-                    path_dbg |= 0x00000100
-                    if decoded != case.exp_message:
-                        path_dbg |= 0x00000040
-                        errors.append(i)
-                        if len(errors) == 1:
-                            path_dbg |= 0x00000200
-                            print(f"FIRST ERROR, at index {i}")
-                            print(f"exp: {case.exp_message}")
-                            print(f"got: {decoded}")
-                else:
-                    path_dbg |= 0x00000400
-                    if case.exp_payload is None:
-                        path_dbg |= 0x00000800
-                        exp_payload = case.src_message.Payload
-                    else:
-                        path_dbg |= 0x00001000
-                        exp_payload = case.exp_payload
-                    if decoded.Payload != exp_payload:
-                        path_dbg |= 0x00000400
-                        errors.append(i)
-                        if len(errors) == 1:
-                            path_dbg |= 0x00002000
-                            print(f"FIRST ERROR, at index {i}")
-                            print(f"exp: {case.exp_payload}")
-                            print(f"got: {decoded.Payload}")
-        # print(f"{decoded.message_type():50s}: path:0x{path_dbg:08X}  {len(errors) == old_len}")    # noqa: ERA001
-    if errors:
-        raise ValueError(f"ERROR. Got codec matching errors at indices {errors}")
-
-
-def test_decoder_simple():
+def test_decoder_simple() -> None:
     child_codec = ChildMQTTCodec()
     parent_codec = ParentMQTTCodec()
-
     assert_encode_decode(child_codec, parent_codec, child_to_parent_messages())
-
     assert_encode_decode(parent_codec, child_codec, parent_to_child_messages())
