@@ -25,7 +25,7 @@ except ImportError:
 
 
 package = "gwproto"
-python_versions = ["3.11", "3.10"]
+python_versions = ["3.11"]
 
 nox.needs_version = ">= 2021.6.6"
 nox.options.sessions = (
