@@ -1,3 +1,5 @@
+# ruff: noqa: ANN401
+
 import uuid
 from typing import Any, Literal
 
@@ -17,9 +19,7 @@ class Ping(BaseModel):
 
 
 class PingMessage(Message[Ping]):
-    def __init__(self, **data: Any):
-        if "AckRequired" not in data:
-            data["AckRequired"] = True
-        if "Payload" not in data:
-            data["Payload"] = Ping()
-        super().__init__(**data)
+    def __init__(self, *, AckRequired: bool = True, **kwargs: Any) -> None:  # noqa: N803
+        if "Payload" not in kwargs:
+            kwargs["Payload"] = Ping()
+        super().__init__(AckRequired=AckRequired, **kwargs)
